feat: migrate sw-meteor-entity-data-table to TypeScript and update styling
