--- src/Administration/Resources/app/administration/src/app/component/entity/sw-meteor-entity-data-table/sw-meteor-entity-data-table.scss
@@ -1,106 +1,0 @@
-/**
- * @sw-package framework
- */
-
-.sw-meteor-entity-data-table.mt-card {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    max-width: none;
-    height: 100%;
-    min-height: 0;
-    margin: 0;
-
-    .mt-card__toolbar {
-        flex: 0 0 auto;
-    }
-
-    .mt-card__toolbar:has(.mt-data-table__toolbar:empty) {
-        display: none;
-    }
-
-    .mt-card__content {
-        display: flex;
-        flex: 1 1 auto;
-        flex-direction: column;
-        min-height: 0;
-    }
-
-    .mt-data-table__table-wrapper {
-        flex: 1 1 auto;
-        min-height: 0;
-    }
-
-    .mt-data-table__footer-inset {
-        flex: 0 0 auto;
-    }
-}
-
-.sw-meteor-entity-data-table {
-    .sw-meteor-entity-data-table__cell {
-        display: flex;
-        align-items: center;
-        gap: 12px;
-        min-width: 0;
-    }
-
-    .sw-meteor-entity-data-table__preview {
-        display: inline-flex;
-        flex: 0 0 auto;
-        align-items: center;
-    }
-
-    .sw-meteor-entity-data-table__column-value {
-        min-width: 0;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        white-space: nowrap;
-    }
-
-    .mt-data-table__table-wrapper-table-row:hover,
-    .mt-data-table__table-wrapper-table-row:focus-within {
-        .sw-meteor-entity-data-table-inline-edit-cell__start {
-            opacity: 1;
-        }
-    }
-}
-
-.sw-meteor-entity-data-table-inline-edit-cell {
-    display: inline-flex;
-    align-items: center;
-    gap: 6px;
-    min-width: 0;
-    max-width: 100%;
-
-    .sw-meteor-entity-data-table-inline-edit-cell__value {
-        min-width: 0;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        white-space: nowrap;
-    }
-
-    .sw-meteor-entity-data-table-inline-edit-cell__start {
-        display: inline-flex;
-        align-items: center;
-        justify-content: center;
-        width: 24px;
-        height: 24px;
-        padding: 0;
-        color: var(--color-icon-secondary-default, #52667a);
-        cursor: pointer;
-        background: transparent;
-        border: 0;
-        opacity: 0;
-        transition: opacity 0.12s ease-in-out;
-
-        &:focus-visible {
-            opacity: 1;
-            outline: 2px solid var(--color-border-brand-default, #189eff);
-            outline-offset: 2px;
-        }
-    }
-
-    .sw-meteor-entity-data-table-inline-edit-cell__input {
-        min-width: 120px;
-    }
-}
